refactor: rename staking tools and files for better naming
- Rename staking-info.ts to staking.ts
- Rename staking-info.test.ts to staking.test.ts
- Rename get-staked-amount-user to get-staked-balance
- Rename get-total-staked-amount-user to get-total-staked
- Update all imports and references
- Update documentation with new tool names
- Maintain functionality while improving naming consistency
