+++ src/index.ts
@@ -2,7 +2,7 @@ import { McpServer } from '@modelcontextprotocol/sdk/server/mcp.js';
 import { StdioServerTransport } from '@modelcontextprotocol/sdk/server/stdio.js';
 import { registerApproveTools } from './tools/approve.js';
 import { registerStakeTools } from './tools/stake.js';
-import { registerStakingInfoTools } from './tools/staking-info.js';
+import { registerStakingInfoTools } from './tools/staking.js';
 import { registerTokenTools } from './tools/token.js';
 import { registerTONCommands } from './tools/ton.js';
 import { registerUnstakeTools } from './tools/unstake.js';
